pyproject.toml
[project]
name = "dbt-warehouse"
version = "0.1.0"
description = ""
authors = [
    {name = "LeonArmster",email = "leo.armster1994@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.8"
dependencies = [
    "pandas (>=2.3.2,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "pg8000 (>=1.31.4,<2.0.0)",
    "dbt-postgres (>=1.9.1,<2.0.0)",
    "streamlit (>=1.49.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
